requirements.txt
# Core dependencies for the Bluesky monitoring project
atproto>=0.0.46

# Flask and web dependencies
flask>=3.0.0
flask-socketio>=5.3.0

# Database dependencies
mysqlclient>=2.2.0
PyMySQL>=1.1.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Utility dependencies
python-dotenv>=1.0.0
requests>=2.31.0

# Development and analysis
jupyter>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0s
seaborn>=0.12.0

# Optional: for better performance
psutil>=5.9.0

# Timezone support
pytz>=2023.3

#flask 

mysql-connector-python==8.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
